task_tracker/pyproject.toml
[tool.poetry]
name = "task-tracker"
version = "0.1.0"
description = "Task Tracker service"
authors = ["Denis Kirisov"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
django-oauth-toolkit = "^2.0.0"
requests = "^2.27.1"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
